test(event): migrate inline tests to integration tests
- Move 173 tests from src/event/*.rs to tests/event_tests.rs
- Keep tests accessing private internals inline:
  - reader.rs: tests using private tick_rate field and convert_key_event fn
  - gesture.rs: test using private calculate_swipe_direction fn
- Convert imports from use super::* to use revue::event::*
